Fix hanging slicer
Two problems were identified. First, file_slicer_t was not
handling pread returning 0 (EOF) and going into an endless loop.

Second, bgzf_slicer_iterator_t was attempting to slice blocks
past the end of the file, which triggered the first problem.

Conceptually it would be nice if we could just avoid reading a
block until the iterator is dereferenced, but that won't work
because we need to read the block in order to know where the next
block is. So the approach taken was simply to make
load_current_block a safe operation that checks to make sure it's
within the file.